Redesign TakeNShots.db as several independent processes
Instead of the flowchart style of the previous version, write the
TakeNShots process database to be closer to how I understand EPICS style.

This breaks the flow into separate processes that trigger based on
conditions specified by calcout records based on TriggerGeneration:Status
and a few other records.

So now the db is separated into
* Watching TriggerGeneration:Status to change to Preparing, which starts
  camera acquisition.
* Checking cameras acquiring, and NumShots and Frequency correct, which
  changes Preparing to Armed.
* Watching TriggerGeneration:Status to change to Idle, which stops
  camera acquisition.