Switch Redis and Endpoint environment variables from .env to os.environ
Instead of using dotenv to pick up IMAGE_BACKEND_ENDPOINT_URL and
IMAGE_BACKEND_REDIS_*, use OS-level environment variables.

This change is made so that the Dockerfile can pass the necessary
environment variables to the application, through build args.

This means that in non-container use, the variables should be set in the
OS-level environment variables.