image_uploader: use caput with char array to set LastAnalyzedShotID
The Win IOC was not seeing the LastAnalyzedShotID set by the
image_uploader client possibly because it uses Channel Access
whereas the image_uploader used pvAccess. But definitely because
the shot_id had well over the max allowed characters, 40.

With the $ notation, this instead sends the string as a char array,
and the receiver needs to convert this back into string.