image_uploader: use caput with char array to set LastAnalyzedShotID
The Win IOC was not seeing the LastAnalyzedShotID set by the
image_uploader client possibly because it uses Channel Access
whereas the image_uploader used pvAccess. But definitely because
the shot_id had well over the max allowed characters, 40.

With the $ notation, this instead sends the string as a char array,
and the receiver needs to convert this back into string.

manually cherry-picked from
30c7137b72789c2045f2c904c31f4c1107a7c3a0

Also handle case where last_analyzed_pv_name cannot be found on the network.