analysis links: change symlink to be relative
Different machines will have the data folder mounted differently, so the symlink that was
created before was to an absolute path on the creating machine, i.e. however it's mounted
on the creating machine.

This commit changes the symlink to point to a shot folder relative to the symlink's parent,
(so it has several "../" to get back up to data from down in the
analysis/year/session/scan/burst/ folder.