Use dialect TIMESTAMP instead of SQLAlchemy DateTime for microsecond precision in burst and shot timestamps
The default for MariaDB is 0-length microsecond precision. This commit
changes the Column spec for Shot.timestamp and Burst.timestamp from
sqlalchemy.types.DateTime to sqlalchemy.dialects.mysql.TIMESTAMP(6) to
use microsecond precision.

I pick mysql.TIMESTAMP instead of mysql.DATETIME because it takes up
half the space specifically in the measurement table, at the expense of the Y2038 problem. But a database isn't a permanent solution anyway.