Set load=False in SQLAlchemySession.merge
This is unnecessary because we don't expect any changes upstream since
last loading the Shot and Variable rows.

It was also a possible reason for the error:
"lock wait timeout exceeded" "raised as a result of Query-invoked
autoflush; consider using a session.no_autoflush block if this flush is
occurring prematurely"

which didn't seem to be solved by using scoped_session.