Create Shots as Measurements for them come in, rather than preallocating
Previously, in the prepare_burst() step, the Burst and associated Shots would
be written to the database, and measurements would reference those Shots.

If a burst is aborted, that leaves a lot of unnecessary shots. And the UI also
allows a burst of infinite number of shots, so shots cannot and should not
be preallocated.

This commit
* Creates a burst.shot_directory to replace the burst.shots list. It's a dict
  instead of a list because it's not guaranteed that all previous shots have
  been created
* Creates Shot during scalar_pv_callback() if it's not yet in the directory
* Removes ScalarSaveData NamedTuple, because a Measurement will do
  just as well, possibly better.
* Removes shots relationship from Burst table object, since it's no longer
  used