Register additional attributes in children of orm.table classes
I've used the Burst, Shot, and Variable instances, which are mapped tables,
to hold attributes of bursts, shots and variables not related to table fields,
(such as Variable.pv to hold the epics PV object, Burst.scalars_saved_tracker
to hold the tracker object for a burst) to keep things organized. But the type
checker does not recognize these.

So this commit registers those attributes by subclassing the table-mapped
classes and adding attributes, dataclass-style.