Add threadsafe method to create new shot
Several callbacks, i.e. threads, check to see if a shot is in burst.shot_directory
and create it if it doesn't exist.

This commit adds a create_new_shot method that uses a threading.Lock to
make sure the shot isn't created multiple times.

Also corrects getting shot and shot timestamp in image_pv_callback.