Don't hold SQLAlchemySession open in ScalarSaveThread
Instead of opening and closing an SQLAlchemySession at the start and end
of ScalarSaveThread.run(), open and close the session only for committing
measurements.

This attempts to solve the issue that Scan and Session objects were associated
with a SQLAlchemySession - presumably the ScalarSaveThread one - so it
wasn't possible to use them in writing the burst in prepare_burst().

There are probably ways to solve this differently/better. For example, don't
write the burst in prepare_burst() but have ScalarSaveThread do it along with
the measurement. Or probably: some permutation of merging.