Get Session, Scan, Burst timestamps from UI instead of generating them.
Previously, the data_uploader generated timestamps with datetime.now()
whenever the Session title or Scan number changed, or when the Burst
status went to Preparing.

In this change, the UI is responsible for generating timestamps, and the
data_uploader monitors these timestamp PVs and creates Sessions, Scans,
and Bursts with these timestamps.

Shot timestamps are still inferred from Burst timestamp and repetition
rate.