Add Timer counter and and check for only one counter during Check
As indicated, thus far we're not accounting for the fact that the
Preparing event after the timeout period could be a different
one.

This commit adds a running count of the number of active timers,
and only alerts a timeout if only one timer is remaining.

Unfortunately this doesn't work because the seq record blocks,
i.e. new Preparing events don't trigger new timers until the first
timer expires.