Support monitoring image PVs over Channel Access
While PVAccess is preferable, some image PVs are only available over
Channel Access. This commit first checks if the image PV is available
over PVA and adds a PVA monitor if so. If not, it checks channel access
and adds a CA monitor.

Adds a CA callback which just adapts the arguments for the PVA callback.