Convert async callbacks into sync callbacks for CA monitors
Add a make_sync_callback function to turn the async callback coroutines
into something that can be passed as a PV monitor callback.

A slightly complicating factor is that PV monitor callbacks are run in
a separate thread (each PV has one, I believe), which doesn't have access
to the event loop. Thus the event loop needs to be bound and passed to
the make_sync_callback function to be used in a run_coroutine_threadsafe
call.