When creating a new Session, sqlalchemy.merge it if it exists in the db
If a UI Session is created with a timestamp that already exists in the
measurement db, a primary key error is raised when trying to add
Measurement objects with that UI Session.

This can happen if the data_uploader is restarted while the UI continues
in the same Session.

Using a SQLAlchemy session merge should bind the Session object to the
existing Session row in the measurement db if it exists.