Put last_analyzed_shot_id_pvs in their own dict
The last_analyzed_shot_id PVs are now related to Instruments, i.e.
sets of devices that are jointly analyzed by the image backend and return
one result. There are no Instrument objects in the DataUploader (yet?)
so just hold these PVs in a dict[InstrumentName, PV]

Also update the callback to take an instrument name.