WIP: Incorporate ICT by adding db records to picoApp
This is the easiest way to incorporate the ICT into the UT3 EPICS control system,
given the PS3000 driver that is included in the picoApp.

This will simply be a process database included in the picoApp that contains records
representing the ICT waveform and integrated charge, which simply use the records
from the PicoScope 3000 in the same IOC.

The alternative would be to build the PS3000 app as a library, and write a separate
ICT IOC that includes this library.