Add 2ms wait after set commands to prevent reading back too fast
The Gentec Integra can handle ASCII commands up to 900 Hz. The EPICS
records that call the set functions usually have a FLNK to the RBV records.
The IOC often gets timeouts on the "get" function calls presumably
because they came too fast, and in addition it appears that the "set"
value isn't properly written either.