Put delays between Set and RBV in process db rather than protocol
To prevent concurrent commands, a delay was put at the end of "set"
protocol functions before the "get" command was called (by forward link)
to get the read-back value (RBV).

This isn't really appropriate; the protocol shouldn't be responsible for
preventing simultaneous commands, and other applications may not have
this issue but are still stuck with delays after Set commands.

This commit puts the delays in the process database by forward-linking
"set" records to a Delay_* seq record that calls the RBV record after
a delay.

Also adds a record to set this delay.